requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
EasyProcess==1.1
entrypoint2==1.1
frozenlist==1.4.0
func-timeout==4.3.5
hydra-core==1.3.2
idna==3.4
jeepney==0.8.0
lxml==4.9.3
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
omegaconf==2.3.0
openai==0.28.1
packaging==23.2
Pillow==9.5.0
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-pptx==0.6.22
python3-xlib==0.15
pytweening==1.0.7
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
termcolor==2.3.0
tiktoken==0.5.1
tqdm==4.66.1
urllib3==2.0.6
XlsxWriter==3.1.9
yarl==1.9.2
func-timeout
google-generativeai
